Lots and lots of changes:
1.  The random number generator has been moved into the xylose library.  The old
Crappy generator was implemented in the newer generic framework and the
testsuite completed successfully using this generator.  In the near future,
we'll change a template parameter in the bottom of rand.hpp to switch to the
new, more glorious, Kiss generator.
    Because the new generator classes generally have unsigned state vectors, I
had to change the cell packing.
Therefore:  RESTART FILES BUILT PRIOR TO THIS VERSION WILL NOT WORK ANY MORE!!!!
Note again: RESTART FILES BUILT PRIOR TO THIS VERSION WILL NOT WORK ANY MORE!!!!

2.  The xylose library has been overhauled in structure and in content.
    a.  There are several pieces that depended on olson-tools.  These pieces
    have been extracted out of olson-tools and inserted into the xylose library.
    b.  Traincars have been renamed 'segmented_vector' and moved into the xylose
    library.
If you have new stuff for the xylose library, please remember to push it back into
the mainline version.  The mainline version will retain build configuration
files for both CMake and BJam, with CMake still being the officially supported
build system.